Update pokemon.json
Both words Rose and Pink are translated both as Rosa in Spanish. And as pink is the widest meaning of them, making it the first one that come to mind when you read that word is better and more understable to use FLOR word that even having a less precise meaning than Rosa is less confusing about the meaning and other uses of the word.